add missing comments
